move data classes to utils
